newer versions of paper and appendix
You should have an earlier version of this paper I think. It is mainly about CA though it also talks about Harderian gland (HG) and other data. It will soon be submitted to the journal Radiation Research but still needs work. It contains an overall survey of the URAP project's big picture , written in a slightly more mathematician friendly form than most of the key published references. Appendix A5 contains some actual math but may not be needed this semester.