Add files via upload
We need to understand the new data Hada kindly sent. Analyzing it is a really good opportunity for us. To make sure we understand her data sheets and some theory within them, we should reproduce the results of Hada2014 Table 2 as much as we can; then we will know how to ask her the right questions (she will not know the theory so we have to translate into language she will understand or can ask her theoretician-statistician, Cucinotta or George or someone). This .xls is using some of her old data sheets she sent in response to a question about 1 item in Table 2 of Hada2014. I worked on the old data sheets for a long time and concluded the following. 1. We now know how to reproduce "simple frequencies" in Table 2 from the old data sheets, apart from a few rows (those at zero dose and some which refer to "split-dose" experiments).  2. I have an approximation to the +- values, but when I use it on results that combine data from all rows that have the same dose it is not quite the right approximation. 3. The item I asked her about is, as I thought, misprinted in Hada2014 and we need to correct it in all our scripts. I doubt it will make any perceptible difference to any of our conclusions, just tiny numerical changes, but we should spot check. To spot check we will make an approximate correction now to get the value within a factor of 2. Then Julien and I will clear up the the factors of 2, by calculations, studying some stat. and if necessary asking Hada.

The temporary approximate correction is the following. In the data frame big_df in row 34  (for Fe600 at dose 0.06) the error value 0.0002 should be replaced by 0.0015, 30 times as big. Please make that correction and do some spot checks; if you get stuck we will discuss at our meeting 
.0600
0.0174
26
1075
175
0.0002
Fe600