Create template_for_calculating_Katz.Rlate.R
This gives the equations which show how to add a Z.b column to the data frame for PBL. The data used is for a completely different set of experiments (on mouse tumors) so ignore it. But the equations used are the same so one can use them for the CA data frame.