Finished graphs.R
For the 6-ion models, I added the 10-dose version and 10-dose no covariance 6-ion version. Added titles and subtitles for each plot. The 10-dose no-covariance Monte Carlo data is now available in folder as well (so a total of 6 models, 4 csv files as the fewer-dosage-points models do not require a separate Monte Carlo)

Added the no-covariance model to the  Rmd file. Minor improvements of the Rmd file.