Add files via upload
Since we never use 0 dose data in the regression calculations I rearranged input so that we never input 0 dose data into the data frame 

I added a 2 parameter model due to Ham, who is working on a SURF summer project which includes statistician's methodology for "parsimonious" models, derived systematically from a multi-parameter model, which use as few adjustable parameters as possible. We should probably carry his methods to their logical conclusion, ending up with the simplest of all models, a TE-only model which is LNT (linear no-threshold) -- a straight line through the origin (apart from the 0.00071 background prevalence) with only the slope adjustable. Remarkably, this is the oldest of radiobiology models for high-charge, high energy (HZE) isotopes. It is still often used. But that is just an amusing sidelight. I think our main job is done and if the Monte-Carlo can be reduced to a few minutes so much the better.