New Synergy and MonteCarlo.R
The Monte Carlo for 3para looks wrong. I will look into it. But after adjusting the negative values of sig0 and dropping negative kap, the MonteCarlo function runs everytime now. See documentation in the files on what to pass into the functions.

AIC, BIC and CV are in Synergy.R

integer values with names "negative_..para_.._.." indicate the number of negative values replaced/dropped in the first 500 draws.